+++ src/contribution.rs
@@ -6,13 +6,13 @@ use crate::book::Book;
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub(crate) struct Contribution {
     author: Option<Author>,
-    author_id: u32,
+    author_id: u64,
     book: Option<Book>,
-    contributable_id: u32,
+    contributable_id: u64,
     contributable_type: ContributableType,
     contribution: Option<ContributionRole>,
     created_at: PlainDateTime,
-    id: u32,
+    id: u64,
     updated_at: PlainDateTime,
 }
