+++ src/reading_format.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub(crate) struct ReadingFormat {
-    id: u32,
+    id: u64,
     // todo!
     format: String,
 }
